Updated test to reflect the recent structural changes
